Making Samza-Sql-Shell commands pluggable by adding CommandHandler
