+++ samza-sql-shell/conf/shell-defaults.conf
@@ -22,6 +22,7 @@
 # Example:
 shell.executor = org.apache.samza.sql.client.impl.SamzaExecutor
 samza.sql.output = compact
+# shell.commandhandlers = org.apache.samza.CustomeCommandHandler1, com.linkedin.CustomeCommandHandler2
 # samza.sql.system.kafka.address = localhost:2181
 # samza.sql.relSchemaProvider.config.schemaDir = /tmp/schemas/
 # samza.sql.ioResolver = config
