+++ samza-sql-shell/src/main/java/org/apache/samza/sql/client/cli/CliConstants.java
@@ -22,7 +22,7 @@ package org.apache.samza.sql.client.cli;
 /**
  * Constant definitions for the shell.
  */
-class CliConstants {
+public class CliConstants {
   public static final String APP_NAME = "Samza SQL Shell";
   public static final String WINDOW_TITLE = "Samza SQL Shell";
   public static final String PROMPT_1ST = "Samza SQL";
@@ -32,6 +32,7 @@ class CliConstants {
   public static final String CONFIG_SHELL_PREFIX = "shell.";
   // Specifies the executor used by the shell
   public static final String CONFIG_EXECUTOR = "shell.executor";
+  public static final String CONFIG_COMMAND_HANDLER = "shell.commandhandlers";
   public static final String DEFAULT_EXECUTOR_CLASS = "org.apache.samza.sql.client.impl.SamzaExecutor";
 
   public static final String WELCOME_MESSAGE;
